requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.3.4
langchain-community==0.3.3
qdrant-client==1.11.3
sentence-transformers==3.1.1
transformers==4.44.2
datasets==3.0.1
ragas==0.1.10
python-dotenv==1.0.1
docling==2.4.0
streamlit==1.39.0
requests==2.32.3